with_gen_ai/requirements.txt
python-dotenv>=1.1.1
openai>=2.2.0
streamlit>=1.50.0